requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
sqlalchemy==1.4.39
pydantic>=1.6.2,<2.0.0
python-dotenv==1.0.1
PyMySQL==1.1.1
passlib==1.7.4
python-jose==3.3.0